Renaming file and test name
